TpetraSegregatedLinearSystem: preparing the linear system/linear solver files (see issue #249)
A bunch of things are going on in this PR:
  1) cleaning a bit TpetraLinearSolver to collect free function into a helper header that will be used in TpetraSegregatedLinearSolver
  2) Adding types in LinearSolverTypes to avoid adding extra templates to functions
  3) Adding a new option in LinearSystem to create a TpetraSegregatedLinearSystem so far it creates a regular TpetraLinearSystem but this will change in the next PR.
