+++ src/LinearSolverConfig.C
@@ -56,7 +56,7 @@ TpetraLinearSolverConfig::load(const YAML::Node & node)
   if (output_level > 0)
   {
     params_->set("Verbosity", Belos::Errors + Belos::Warnings + Belos::StatusTestDetails);
-    params_->set("Output Style",Belos::Brief); 
+    params_->set("Output Style",Belos::Brief);
   }
 
   params_->set("Output Frequency", output_level);
@@ -99,11 +99,12 @@ TpetraLinearSolverConfig::load(const YAML::Node & node)
     throw std::runtime_error("invalid linear solver preconditioner specified ");
   }
 
-  get_if_present(node, "write_matrix_files", writeMatrixFiles_, writeMatrixFiles_);
-  get_if_present(node, "summarize_muelu_timer", summarizeMueluTimer_, summarizeMueluTimer_);
+  get_if_present(node, "write_matrix_files",       writeMatrixFiles_,        writeMatrixFiles_);
+  get_if_present(node, "summarize_muelu_timer",    summarizeMueluTimer_,     summarizeMueluTimer_);
 
   get_if_present(node, "recompute_preconditioner", recomputePreconditioner_, recomputePreconditioner_);
   get_if_present(node, "reuse_preconditioner",     reusePreconditioner_,     reusePreconditioner_);
+  get_if_present(node, "segregated_solver",        useSegregatedSolver_,     useSegregatedSolver_);
 
 }
